increase events to 3000, reduce the window size to 400
the reduction of window size makes it run for long time, i would still suggest to use 1000, but it leaded to only around 15 windows, which is very weird, as the REH with 2000 events simulated in "ver3" has 175 windows with minimum window size 1000. I am not sure is that because I only used `inertia` here, but in ver3 I put `inertia` and `transitivity` into the model.

On the other hand, I changed the thresholds for CPs detection, it doesnt solve the problem of having many CPs of statistics scale in constant parameter data(no CP data).

The other weird thing, even I used constant parameter data(no CP data), but when I fit MW on it, I still got fluctuated parameters cross windows; Last time in ver3 the CPs were detected in the first 30 windows and the last 30 windows, and our conclusion is that it may be because in the beginning and in the end, human interactions are not stable, but this time in ver4 even the windows in the middle got detected CPs.