Solve the problem of detecting CPs in non-CP REH on parameter scale
Make the window size smaller to 1/10 of the events, the CPs of the parameters along the window werent detected anymore. So the main problem before may be is the size of windows in DDMW were set too large.

However, the REH with CP still has some problem that the function inside the REH simulation doesnt work properly, the parameters along the window also showed no CP, but it should had one.